refactor(web-mobile): split admin_competition.jsx into cohesive section modules
admin_competition.jsx was 1723 lines — high merge-conflict surface. Split the
section components into four mutually-independent sibling modules, leaving a
thin entry that holds the AdminCompetition shell and re-exports the full public
surface. Code moved verbatim; rendered UI is unchanged.

New modules (each consumes only globals from earlier admin_* files):
- admin_competition_overview.jsx  — AdminCompOverview, FightingSpiritAwardsEditor
- admin_competition_settings.jsx  — AdminSettings (+ formatCompMinutes helper)
- admin_competition_bracket.jsx   — AdminBracket, RunningMatchPanel, scoreboard helpers
- admin_competition_swiss.jsx     — AdminSwissRounds + Swiss round predicates

The thin admin_competition.jsx entry evaluates every section module (via the
helper re-exports + a side-effect import for overview), aliases their window.*
component globals for the AdminCompetition render tree, and re-exports all eight
pure helpers so existing `import { … } from './admin_competition.jsx'` test
sites keep resolving. index.html loads the four modules (leaf-first) before the
entry; esbuild auto-globs the new .jsx. Cache-bust bumped to ?v=6.

Tier 3 of the safety net: the AdminSettings source-introspection tests now read
admin_competition_settings.jsx, with an explicit non-vacuity guard
(expect(src).toContain('function AdminSettings')) so a wrong path fails loudly
rather than passing on an empty regex match.

Verification:
- Safety net green against the split: contract (public surface preserved),
  static export checker (entry→module re-export edges valid), render-smoke
  (all 5 sections mount).
- Browser smoke (real mobile-app): dynamic import of all 5 /dist modules throws
  nothing, all 6 window globals wired, Overview/Settings/Fighting-Spirit render
  with live data, zero new console errors.
- make go/test green, JS lint 0 warnings, 1671 default + 27 render tests pass.

Closes mp-hpe3.
